+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Operation `orders.create(options)` (nepalez)
 * Operation `orders.find(id)` (nepalez)
+* Operation `files.download` (nepalez)
 
 [Compare v0.0.5...HEAD](https://github.com/nepalez/abbyy-cloud/compare/v0.0.5...HEAD)